requirements.txt
annotated-types==0.7.0 ; python_version == "3.11"
anyio==3.7.1 ; python_version == "3.11"
association-measures==0.3.0 ; python_version == "3.11"
asttokens==3.0.0 ; python_version == "3.11"
bokeh==3.6.2 ; python_version == "3.11"
bottleneck==1.4.0 ; python_version == "3.11"
bqplot==0.12.44 ; python_version == "3.11"
certifi==2024.12.14 ; python_version == "3.11"
charset-normalizer==3.4.1 ; python_version == "3.11"
click==8.1.8 ; python_version == "3.11"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version == "3.11"
comm==0.2.2 ; python_version == "3.11"
contourpy==1.3.1 ; python_version == "3.11"
cwb-ccc==0.12.3 ; python_version == "3.11"
decorator==5.1.1 ; python_version == "3.11"
et-xmlfile==2.0.0 ; python_version == "3.11"
executing==2.1.0 ; python_version == "3.11"
fastapi==0.105.0 ; python_version == "3.11"
ftfy==6.3.1 ; python_version == "3.11"
gast==0.4.0 ; python_version == "3.11"
h11==0.14.0 ; python_version == "3.11"
httptools==0.6.4 ; python_version == "3.11"
humlab-penelope==0.8.2 ; python_version == "3.11"
idna==3.10 ; python_version == "3.11"
ipycytoscape==1.3.3 ; python_version == "3.11"
ipydatagrid==1.4.0 ; python_version == "3.11"
ipyfilechooser==0.6.0 ; python_version == "3.11"
ipython==8.31.0 ; python_version == "3.11"
ipywidgets==8.1.5 ; python_version == "3.11"
jedi==0.19.2 ; python_version == "3.11"
jinja2==3.1.5 ; python_version == "3.11"
joblib==1.4.2 ; python_version == "3.11"
jupyter-bokeh==4.0.5 ; python_version == "3.11"
jupyterlab-widgets==3.0.13 ; python_version == "3.11"
loguru==0.7.3 ; python_version == "3.11"
lxml==5.3.0 ; python_version == "3.11"
markupsafe==3.0.2 ; python_version == "3.11"
matplotlib-inline==0.1.7 ; python_version == "3.11"
more-itertools==10.6.0 ; python_version == "3.11"
msgpack==1.1.0 ; python_version == "3.11"
networkx==3.4.2 ; python_version == "3.11"
nltk==3.9.1 ; python_version == "3.11"
numexpr==2.10.2 ; python_version == "3.11"
numpy==1.26.4 ; python_version == "3.11"
openpyxl==3.1.5 ; python_version == "3.11"
packaging==24.2 ; python_version == "3.11"
pandas==2.2.3 ; python_version == "3.11"
parso==0.8.4 ; python_version == "3.11"
pexpect==4.9.0 ; python_version == "3.11" and sys_platform != "win32" and sys_platform != "emscripten"
pillow==11.1.0 ; python_version == "3.11"
prompt-toolkit==3.0.50 ; python_version == "3.11"
ptyprocess==0.7.0 ; python_version == "3.11" and sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3 ; python_version == "3.11"
py2vega==0.6.1 ; python_version == "3.11"
pyarrow==17.0.0 ; python_version == "3.11"
pydantic-core==2.27.2 ; python_version == "3.11"
pydantic==2.10.5 ; python_version == "3.11"
pygments==2.19.1 ; python_version == "3.11"
python-dateutil==2.9.0.post0 ; python_version == "3.11"
python-dotenv==1.0.1 ; python_version == "3.11"
pytz==2024.2 ; python_version == "3.11"
pyyaml==6.0.2 ; python_version == "3.11"
regex==2023.12.25 ; python_version == "3.11"
requests==2.32.3 ; python_version == "3.11"
scipy==1.12.0 ; python_version == "3.11"
six==1.17.0 ; python_version == "3.11"
smart-open==7.1.0 ; python_version == "3.11"
sniffio==1.3.1 ; python_version == "3.11"
spectate==1.0.1 ; python_version == "3.11"
stack-data==0.6.3 ; python_version == "3.11"
starlette==0.27.0 ; python_version == "3.11"
toml==0.10.2 ; python_version == "3.11"
tornado==6.4.2 ; python_version == "3.11"
tqdm==4.67.1 ; python_version == "3.11"
traitlets==5.14.3 ; python_version == "3.11"
traittypes==0.2.1 ; python_version == "3.11"
trieregex==1.0.0 ; python_version == "3.11"
typing-extensions==4.12.2 ; python_version == "3.11"
tzdata==2025.1 ; python_version == "3.11"
unidecode==1.3.8 ; python_version == "3.11"
urllib3==2.3.0 ; python_version == "3.11"
uvicorn==0.24.0.post1 ; python_version == "3.11"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version == "3.11"
watchfiles==1.0.4 ; python_version == "3.11"
wcwidth==0.2.13 ; python_version == "3.11"
websockets==14.2 ; python_version == "3.11"
wheel==0.45.1 ; python_version == "3.11"
widgetsnbextension==4.0.13 ; python_version == "3.11"
win32-setctime==1.2.0 ; python_version == "3.11" and sys_platform == "win32"
wrapt==1.17.2 ; python_version == "3.11"
xyzservices==2025.1.0 ; python_version == "3.11"
